--- src/server/renderToStream/constants.ts
@@ -1,2 +1,0 @@
-// The default timeout for when we abort react rendering
-export const DEFAULT_TIMEOUT = 20000;
